+++ src/ducks/vurdering/index.js
@@ -1,0 +1,15 @@
+import * as vurderingOperations from './operations';
+import * as vurderingSelectors from './selectors';
+import * as vurderingActions from './actions';
+import * as vurderingTypes from './types';
+
+import vurderingReducers from './reducers';
+
+export {
+  vurderingOperations,
+  vurderingSelectors,
+  vurderingTypes,
+  vurderingActions,
+};
+
+export default vurderingReducers;
