+++ src/ducks/vurdering/types.js
@@ -1,0 +1,3 @@
+export const OK = 'vurdering/OK';
+export const FEILET = 'vurdering/FEILET';
+export const PENDING = 'vurdering/PENDING';
